refactor: complements the layered architecture of the get all products route
