+++ src/routes/productsRouter.js
@@ -1,10 +1,10 @@
-import express from "express";
+import { Router } from "express";
 
 import validateProduct from "../middlewares/validateProductMiddleware.js";
 
-import { registerProduct, findProduct, getProducts } from "../controllers/productController.js";
+import { registerProduct, findProduct, getProducts } from "../controllers/productsController.js";
 
-const productsRouter = express.Router();
+const productsRouter = Router();
 
 productsRouter.post("/product", validateProduct, registerProduct);
 productsRouter.get("/products", getProducts);
